+++ library/src/main/java/com/evernote/android/job/Job.java
@@ -113,10 +113,22 @@ public abstract class Job {
     @WorkerThread
     protected abstract Result onRunJob(@NonNull Params params);
 
+    /**
+     * This method is intended to be overwritten. It is called once when the job is still running, but was
+     * canceled. This can happen when the system wants to stop the job or if you manually cancel the job
+     * yourself. It's a good indicator to stop your work and maybe retry your job later again. Alternatively,
+     * you can also call {@link #isCanceled()}.
+     *
+     * @see #isCanceled()
+     */
+    protected void onCancel() {
+        // override me
+    }
+
     /*package*/ final Result runJob() {
         try {
             // daily jobs check the requirements manually
-            if (this instanceof DailyJob || meetsRequirements()) {
+            if (this instanceof DailyJob || meetsRequirements(true)) {
                 mResult = onRunJob(getParams());
             } else {
                 mResult = getParams().isPeriodic() ? Result.FAILURE : Result.RESCHEDULE;
@@ -142,8 +154,18 @@ public abstract class Job {
         // override me
     }
 
-    /*package*/ boolean meetsRequirements() {
-        if (!getParams().getRequest().requirementsEnforced()) {
+    /**
+     * Checks all requirements for this job. It's also possible to check all requirements separately
+     * with the corresponding methods.
+     *
+     * @return Whether all set requirements are met.
+     */
+    protected boolean meetsRequirements() {
+        return meetsRequirements(false);
+    }
+
+    /*package*/ boolean meetsRequirements(boolean checkRequirementsEnforced) {
+        if (checkRequirementsEnforced && !getParams().getRequest().requirementsEnforced()) {
             return true;
         }
 
@@ -158,6 +180,15 @@ public abstract class Job {
         if (!isRequirementNetworkTypeMet()) {
             CAT.w("Job requires network to be %s, but was %s", getParams().getRequest().requiredNetworkType(),
                     Device.getNetworkType(getContext()));
+            return false;
+        }
+        if (!isRequirementBatteryNotLowMet()) {
+            CAT.w("Job requires battery not be low, reschedule");
+            return false;
+        }
+
+        if (!isRequirementStorageNotLowMet()) {
+            CAT.w("Job requires storage not be low, reschedule");
             return false;
         }
 
@@ -263,7 +294,10 @@ public abstract class Job {
 
     /*package*/ final void cancel(boolean deleted) {
         if (!isFinished()) {
-            mCanceled = true;
+            if (!mCanceled) {
+                mCanceled = true;
+                onCancel();
+            }
             mDeleted = deleted;
         }
     }